requirements.txt
torch
transformers
datasets
jsonlines
numpy
scipy
huggingface-hub
tqdm
optimum
accelerate
docker
google-cloud-storage
google-cloud-compute